tests/requirements.txt
selenium==4.20.0
pytest==8.2.0
pytest-html==4.1.1
pytest-cov==5.0.0
pytest-xdist==3.5.0
pytest-timeout==2.3.1
webdriver-manager==4.0.1
python-dotenv==1.0.1
Faker==24.9.0
